Rearranging item and some stub macros
