+++ src/module/macros/talent.ts
@@ -1,0 +1,20 @@
+// Import TypeScript modules
+import { LANCER } from "../config";
+import type { LancerActor } from "../actor/lancer-actor";
+import type { LancerTalentMacroData } from "../interfaces";
+import { renderMacroTemplate } from "./render"
+
+const lp = LANCER.log_prefix;
+
+export async function rollTalentMacro(actor: LancerActor, data: LancerTalentMacroData) {
+  if (!actor) return Promise.resolve();
+
+  // Construct the template
+  const templateData = {
+    title: data.talent.name,
+    rank: data.talent.ranks[data.rank],
+    lvl: data.rank,
+  };
+  const template = `systems/${game.system.id}/templates/chat/talent-card.hbs`;
+  return renderMacroTemplate(actor, template, templateData);
+}
